separate the tcp specific parts from the main client request loop
