+++ rodbus/src/client/task.rs
@@ -1,11 +1,8 @@
-use std::net::SocketAddr;
 use std::time::Duration;
 
-use log::{info, warn};
 use tokio::prelude::*;
 use tokio::sync::*;
 
-use crate::client::channel::ReconnectStrategy;
 use crate::client::message::{Request, ServiceRequest};
 use crate::error::*;
 use crate::service::function::ADU;
@@ -18,7 +15,7 @@ use crate::util::frame::{FrameFormatter, FrameHeader, FramedReader, TxId};
 /**
 * We always service requests in a TCP session until one of the following occurs
 */
-enum SessionError {
+pub(crate) enum SessionError {
     // the stream errors or there is an unrecoverable framing issue
     IOError,
     // the mpsc is closed (dropped)  on the sender side
@@ -34,62 +31,24 @@ impl SessionError {
     }
 }
 
-/// Channel loop
-///
-/// This loop handles the request one by one. It serializes the request
-/// and sends it through the socket. It then waits for a response, deserialize
-/// it and sends it back to the oneshot provided by the caller.
-pub struct ChannelTask {
-    addr: SocketAddr,
+pub(crate) struct ClientLoop {
     rx: mpsc::Receiver<Request>,
-    connect_retry: Box<dyn ReconnectStrategy + Send>,
     formatter: MBAPFormatter,
     reader: FramedReader<MBAPParser>,
     tx_id: TxId,
 }
 
-impl ChannelTask {
-    pub fn new(
-        addr: SocketAddr,
-        rx: mpsc::Receiver<Request>,
-        connect_retry: Box<dyn ReconnectStrategy + Send>,
-    ) -> Self {
+impl ClientLoop {
+    pub fn new(rx: mpsc::Receiver<Request>) -> Self {
         Self {
-            addr,
             rx,
             formatter: MBAPFormatter::new(),
-            connect_retry,
             reader: FramedReader::new(MBAPParser::new()),
             tx_id: TxId::default(),
         }
     }
 
-    pub async fn run(&mut self) {
-        // try to connect
-        loop {
-            match tokio::net::TcpStream::connect(self.addr).await {
-                Err(e) => {
-                    warn!("error connecting: {}", e);
-                    let delay = self.connect_retry.next_delay();
-                    if self.fail_requests_for(delay).await.is_err() {
-                        // this occurs when the mpsc is dropped, so the task can exit
-                        return;
-                    }
-                }
-                Ok(stream) => {
-                    info!("connected to: {}", self.addr);
-                    match self.run_session(stream).await {
-                        // the mpsc was closed, end the task
-                        SessionError::Shutdown => return,
-                        // re-establish the connection
-                        SessionError::IOError => {}
-                    }
-                }
-            }
-        }
-    }
-
-    async fn run_session<T>(&mut self, mut io: T) -> SessionError
+    pub async fn run<T>(&mut self, mut io: T) -> SessionError
     where
         T: AsyncRead + AsyncWrite + Unpin,
     {
@@ -233,7 +192,7 @@ impl ChannelTask {
         }
     }
 
-    async fn fail_requests_for(&mut self, duration: Duration) -> Result<(), ()> {
+    pub async fn fail_requests_for(&mut self, duration: Duration) -> Result<(), ()> {
         let deadline = tokio::time::Instant::now() + duration;
 
         loop {
