+++ rodbus/src/tcp/mod.rs
@@ -1,1 +1,2 @@
+pub(crate) mod client;
 pub(crate) mod frame;
